+++ src/function/isEqual.js
@@ -1,4 +1,5 @@
-import _equal from '../_internals/equal'
+import curry from '../function/curry'
+import _equal from '../_internals/_equal'
 
 /**
  * @name isEqual
@@ -6,14 +7,14 @@ import _equal from '../_internals/equal'
  * @category Function
  * @sig a -> b -> Boolean
  * @description Takes and compares two items. Capable of handling cyclical data structures
- * @param  {Any} a First item to compare
- * @param  {Any} b Second item to compare
+ * @param {Any} a First item to compare
+ * @param {Any} b Second item to compare
  * @return {Boolean} Returns the boolean after running our comparison check
  *
  * @example
  * const obj = isEqual({}, {}); // => true
  * const arr = isEqual([], []); // => true
  */
-const isEqual = (a, b) => _equal(a, b, [], [])
+const isEqual = (a, b) => _equal(a, b)
 
-export default isEqual
+export default curry(isEqual)